+++ Changelog.md
@@ -4,8 +4,12 @@
 * use a single state atom in simulator
 * added perfect power factor to math
 * generalized quantum phase estimation and integrated in quantum-period-finding
-* phase estimation handles the number of measurements, reduced the number of circuit creations 
+* phase estimation handles the number of measurements, reduced the number of circuit creations
+* fixed edge cases for quantum phase estimation
 * fixed coprime selection in shor
+* refactored tests for quantum period finding
+* added prime? function
+* added more tests for the math namespace
 * added prime check in shor
 
 ## Version 0.1.0